add click functionality to gui and allow the player to move the pieces
it also appends the window title with the current player color

there is two states, the first click that is the piece you want to
move and the second, that is where you want to place it.

If youclick in a square/spot that has no pieces it will not be lead
to second state since _start_move_piece checks if the position has
a piece within.

In the end of the second state, _stop_move_piece
will call game.move to change the piece and game.move will send
back a message telling if the move is invalid and the message
will be shown on a messagebox.