separate move logic from the state switch logic and is_fist_move flag on piece
now the entire turn logic is on game.play_turn witch checks the movestate
returned by game._check_move_state and process it. If is a valid move then
it calls game._move to update the game.board and then change the turn
or throws an error when a move is invalid to the gui catch and shows to the
player or can just ignore everything if the move is invalid but something that
the player should not be bothered by receiving a message (such as clicking
on a empty spot or the other player piece).

i can create custom exceptions to deal with other types of reactions i want
the gui to have go each kind o move the player tries to do later (for example
instead of ignoring when the player click in a invalid position it can play
a sound or highlight its own  pieces). Therefore i think this strategy of
enums/exceptions is better than let the gui know too much about how the game
works but is not ideal either

with all this said, maybe 'play_turn' is not the best name for the method