add 'check' system for preventing the player of putting a pice when the king is in check
this big commit adds the functionality to prevent the player of moving
pieces when they king is in check and from impeding the king of moving
to squares where it could be in check.

the system is not perfect since it relies on the moving logic of each
piece, that is not complete yet (they aren't aware of pices in-between
the square from where it came from and from where is going). When this
logic is completed the check system is going to work without any further
programming.