add get_pieces_range_diagonal to get the pieces between two positions
this gave me a lot of troubles but in the end it was the first of
the three 'get_pieces_range' i finished the implementation. Now is
possible to finish the bishop's moving logic.

maybe i should reconsider return (x, y, piece) and return only
(x,y, color) since i think the pice will not be really useful