add unit test for the board class
currrently this only test the method get_pieces_range_diagonal. It
relies on the utility functions within utils module to set up the
initial board state and to make result (a list of tuples, which is
the return type from all get_pieces_range_xxx methods) in a more
readable and writable way.