add tests for board.is_in_check and for  Knight
this also renames BoardTest => BoardTests

this also add three lists each with all the moves a knight can
do in one turn. One list is abstract with marks x,y,w,z to be
replaced in the other two list with the knigth (x or y) and to
add a pice where it will move to (again x or y) and also there
is two marks to add the kings (w,z, one of them is ignored in
the two lists build on top of the abstract one since it will be
used to check if a knight is checking the king therefore the
knight desired position is the king)

the tests test for squares where one knight (white or black)
can land, and for the places where it can't land because
there is a friendly piece or a empty square. # Please enter the commit message for your changes. Lines starting