add tests for board.is_square_in_check
I'm still not testing if it works with the black color yet and
neither with the pawn since i'm using the pawn itself as a markation
so i'll leave it to its own test.

I found a bug in the in_check logic because the way the pawn can
normally move is not the same way when it can capture another
piece, and this is not considered when i loop thru all pieces
to test the can_move, so i have to make a special case to the
pawn. And of course, i have to test the cenarios where the pawn
can and can't capture the king