'piece' can be deserialized as '0' for tests purposes
Adds the piece's superclass Piece as a possible
piece to be loaded by utils.py load_board/piece_from_char
functions. This is useful since we can use it to pinpoint
the place i want a given piece to go within the tests
and get use the board functions to simply get the position
from the Piece piece and them remove it instead of manually
counting the rows and columns to get the correct position,
what makes easier to set up a layout and edit it without
having to change the actual test code.

NOTE: the deserialized piece.Piece is always black