add diagonal backward tests for the pawn
it also adds the previously commited tests to the white pawn
(since all testes were made to work with the player that is
on top, which was the white but later was switched to black
to make the queens be in the right squares) and can not move
backwards twice tests

and each test has now it's own layout being load since was getting
messy to figure out which pieces were beng used in each test.
There's a lot of duplicate code but i can deal with it later.

the pawn can't go backwards diagonally even if is to capture a
piece, currently this test fails since the engine is allowing
this illegal movement.