abstract way the the internals from the piece tests into PieceTestBase
this should make the tests easy to read and write. Only the board
layout and the pieces you're manipulating need to be expecified
and the custom asserts from PieceTestBase hides all the shared
piece-get_location/remove/etc logic.

within PieceTestBase class still has duplicate code but we can
fix it later, what matters is that the actual tests are clean.

later i should do something similar to board_test since that is
still a mess.