major refactoring in in_check/king.can_move logic fixes #6
this is even more coupled than before but now all tests pass and
i can make sense of what is happening. The previous commit that
enabled king to eat and not come close to the other king is also
working properly.

that said, we need more tests for the king