implement capture system
with this system now the pieces returns a MoveResult instead of
a bool, the pieces feed the MoveResult with the position and
the piece they captured to the game deal with it.

up to now, the only 'captured' pieces that were removed from
the board were the ones where the piece that is trying to
move landed. With this pack, now the pieces and the game
are aware of the captured (even the ones in the way from a move)
pieces and the game can do something with this information.

maybe the i.a will need it, if not we can use it to implemente a
score system or something like that.

now we need to implement some tests to the game and change the
name from can_move to something more meaningful