prevent moves that put or leave the king in check
the old logic i did before only accounted to prevent from moving any
piece but the king when it is in check, which doesn't make sense
since a move can cancel the check. This fixes that and follow the
rule that that say that you can't do a move if the move put the
king in check neither do a move when the king is already in check
and that do not prevent that.