fix bug that throws not king exception
_move_will_leave_in_check_state throws no king of {current turn
color} found when a piece from the current player is clicked and
then its king is clicked after. It seems the is_position_in_check
(or the method cited prior) did something that lost the king from
the board. _move_will_leave_in_check_state shouldn't do this, i
see the board after the _move and the king was there (the piece
vertically above was not tho).

this is a workaround since this is such a corner case that. If
i return with NO_PIECE_TO_MOVE we wouldn't have to fix it but
since the castling needs to click in two of your pieces didn't
make the check before. Let's hope i'd ever make the moves atomic
since this would fix this and we would be able to get rid from
the hack.