add primitive checkmate/draw checkings
this adds the base-minimum end game logic.W e don't really plan to
suport all kinds of way of ending a game by checkmating or drawing
since is too much of a bother to a side project.

this patch is somewhate hacky as you can see in play_turn and we
need to re organize some stuff and at least get rid of that chained
IF statements.

aside from a showing messagebox, this don't do anything with the
like showing feedback in the board ui or allowing to play again.