remove checkmate logic
for now, our version from the game ends when the king is captured,
which is invalid in the regular chess. We can't avoid this since
we don't have time to implement the full logic yet and we need
to start working on the a.i. If we have time before the presentation
or if i ever fell like working with this after that then i'll
either revert this or start the checkmate/draw logic from scratch.