create method to get all valid moves of a color, except king
this commit adds the last needed functionality to implement the a.i.
It iterate through all pieces of a certain color (a.k.a material),
then get its pseudo moves and validate it. This implementation, as
you already read above, only gets from one color at call, but if
i remember correctly, the implementation of the i.a needs from
both colors so the implementation may change in the future.