implement the i.a partially
first i create the ia.py file where i put all the a.i logic,
which currently is two classless functions. One uses the
get_value from piece to do some math and the other is a
minimax with alpha beta pruning implementation that i adapted
from the link i leaved commented on the top of the file. It
would have made sense if i put this in the game class but i
want a isolate this logic and since it needs so much of the
game methods, i'll propably make it a class with a game
instance.

this patch touches a lot of files. Since i want to see the movement
ocorring and not just the outcome of the player move + a.i then i
divided the code that make the a.i wihin two methods then i called
then after the UI rendered the player's move.