fix #15 bishop/queen/rook move where we allowed them to capture on jump over
this is not dames so the queen can't jump over an capture a piece
that was in the middle of path. How could a have not realized until
now is beyound me.

This patch also adds some checks in the "can not move" tests to
check if it's allowing to jump over (there was none so far).
There still some tests that aren't in all pieces with row/colunm
-wide movement but i think that this will not be a problem.