remove manual move and undo
when the game checks if the a move will leave the current player
to a check state it manually move and undo the move, this was
implemented before the moveresult implemented the necessary
data to make the moves atomic. Now the'll use the default
implementation instead of doing 'by hand'.