fix bug introduced in "fix bug that throws not king exception"
the != is probably because i inverted the logic and didn't notice since this
was never need before since the player moves were input validated and the a.i
do not call it. Anyway, since the method will be called by the a.i then we i
set manually the piece in the landed position since she will be overriten if
the result move is false (we shouldn't call the method if the player can't
move but the way i handle the "will be in check" logic is just messed up).

i add a condition to ignore if the piece were it is landing is the
king so we don't have to go through all the in check logic inside
of its can_move.

tl;dr this patch will prevent the a.i's king and the player's pieces
from being overriten when the a.i checks for all valid moves.