prevent the a.i from crashing when no moves were available
i doubt the a.i had'nt any moves to do because of a stalemate,
some bug(1) prevented it but is nice to prevent the crash anyway.
I did make to a position where it should stalemate the a.i but
it didn't care and continued to move its other pieces. (bug 2)