fix checkmate tgat was broken by promotion
the game was not ending since the move was made after the promotion/
end game checks, now i do the move without the promotion and then
undo to promote correctly. This is a huge hack but it works.

Something to consider is that after the promotion it could checkmate
but the engine didn't check for that.