fix checkmate message not showing or showing the wrong player
this fixes a bug where the checkmate message was not shown. That
was caused by my 'fix' for preventing the promotion dialog to
appear if the other player was checkmated. Since i moved the check
to before the move, the move was computed after the checks for
end game. Now we will promote always, and then check for end game.

this also fixes wrong player being displayed in the win message.